# Conflicts:
#	emovie_server/src/main/java/com/emovie/dao/MovieDao.java
#	emovie_server/src/main/java/com/emovie/service/IMovieService.java
#	emovie_server/src/main/java/com/emovie/service/impl/MovieServiceImpl.java
#	emovie_server/target/test-classes/com/emovie/EmovieServerApplicationTests.class

+++ emovie_server/src/main/java/com/emovie/entity/Movie.java
@@ -75,7 +75,7 @@ public class Movie {
 	 * table type:decimal(8)
 	 * table comment:null
 	 */
-	private Double popularity;
+	private Integer popularity;
 
 	/**
 	 * table name:poster_path
@@ -131,7 +131,7 @@ public class Movie {
 	 * table type:decimal(8)
 	 * table comment:null
 	 */
-	private Double voteAverage;
+	private Integer voteAverage;
 
 	/**
 	 * table name:vote_count
